Saving information
This app will store information even the app is killed. If we again start the app it will show us the stored data.
Persist data across user sessions, even if app is killed and restarted, or device is rebooted.
Data that should be remembered across sessions, such as a user's preferred settings or their game score.
Common use is to store user preferences.